Fix some type errors

+++ qubrabench/bench/stats.py
@@ -5,6 +5,6 @@ from dataclasses import dataclass
 class QueryStats:
     classical_control_method_calls: int = 0
     classical_actual_queries: int = 0
-    classical_expected_queries: int = 0
-    quantum_expected_classical_queries: int = 0
-    quantum_expected_quantum_queries: int = 0
+    classical_expected_queries: float = 0
+    quantum_expected_classical_queries: float = 0
+    quantum_expected_quantum_queries: float = 0
